Minor Speedup for __output__ setting for new __setattr__ behavior.
Previous behavior made the __output__ flag go through really expensive __setattr__ (which is optimized for row data, not class attributes which have to go through an expensive error handeling)

This is usually fine as inheritance is discourage and class attributes are rarely added at runtime. Though since __output__ is a critical part of the row as an attribute, it had to go through this expensive check for no real reason.